fix octopus file naming convention
