+++ packages/octopus-fig/src/services/exporters/node/debug-exporter.ts
@@ -3,16 +3,10 @@ import path from 'path'
 
 import { detachPromiseControls } from '@opendesign/octopus-common/dist/utils/async'
 import { timestamp } from '@opendesign/octopus-common/dist/utils/timestamp'
+import kebabCase from 'lodash/kebabCase'
 import { v4 as uuidv4 } from 'uuid'
 
-import {
-  getOctopusFileName,
-  getPreviewFileName,
-  getSourceFileName,
-  IMAGES_DIR_NAME,
-  IMAGE_EXTNAME,
-  MANIFEST_NAME,
-} from '../../../utils/exporter'
+import { MANIFEST_NAME } from '../../../utils/const'
 import { makeDir, saveFile } from '../../../utils/files'
 import { stringify } from '../../../utils/misc'
 
@@ -28,6 +22,21 @@ type DebugExporterOptions = {
   tempDir: string
   /** Unique ID of the exported design. If not given, will be generated by UUIDv4. Exporter outputs will be generated into `tempDir/id` folder. */
   designId?: string
+}
+
+export const IMAGES_DIR_NAME = 'images'
+export const IMAGE_EXTNAME = '.png'
+
+export function getOctopusFileName(id: string): string {
+  return `${kebabCase(id)}-octopus.json`
+}
+
+export function getPreviewFileName(id: string): string {
+  return `${kebabCase(id)}-preview${IMAGE_EXTNAME}`
+}
+
+export function getSourceFileName(id: string): string {
+  return `${kebabCase(id)}-source.json`
 }
 
 /**
