+++ packages/octopus-fig/src/utils/const.ts
@@ -1,0 +1,1 @@
+export const MANIFEST_NAME = 'octopus-manifest.json'
