+++ packages/octopus-fig/test/integration/assets/79r1c5joN7rUmT2StGPPGA/expected/octopus-manifest.json
@@ -1,5 +1,5 @@
 {
-  "version": "3.0.0-rc.18",
+  "version": "3.0.0-rc.19",
   "origin": {
     "name": "FIGMA",
     "version": "0"
