+++ packages/octopus-fig/test/integration/assets/ROzbQpuLmg8uDFEFb6OUtZ/expected/octopus-2-3.json
@@ -1,22 +1,22 @@
 {
-  "id": "1-3",
+  "id": "2-3",
   "type": "ARTBOARD",
-  "version": "3.0.0-rc.6",
+  "version": "3.0.0-rc.19",
   "dimensions": {
-    "width": 300,
-    "height": 100
+    "width": 800,
+    "height": 200
   },
   "content": {
-    "id": "1-3-Background",
-    "name": "BUTTON COMP",
+    "id": "2-3-Background",
+    "name": "Text Frame ",
     "type": "MASK_GROUP",
     "visible": true,
     "opacity": 1,
     "blendMode": "PASS_THROUGH",
     "maskBasis": "BODY_EMBED",
     "mask": {
-      "id": "1-3-BackgroundMask",
-      "name": "BUTTON COMP",
+      "id": "2-3-BackgroundMask",
+      "name": "Text Frame ",
       "type": "SHAPE",
       "visible": true,
       "opacity": 1,
@@ -44,11 +44,10 @@
           ],
           "rectangle": {
             "x0": 0,
-            "x1": 300,
+            "x1": 800,
             "y0": 0,
-            "y1": 100
-          },
-          "cornerRadius": 30
+            "y1": 200
+          }
         },
         "fillRule": "NON_ZERO",
         "fills": [
@@ -57,8 +56,8 @@
             "visible": true,
             "blendMode": "NORMAL",
             "color": {
-              "r": 0.3958,
-              "g": 0.565,
+              "r": 1,
+              "g": 1,
               "b": 1,
               "a": 1
             }
@@ -78,7 +77,7 @@
     "layers": [
       {
         "id": "2-4",
-        "name": "Hello",
+        "name": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
         "type": "TEXT",
         "visible": true,
         "opacity": 1,
@@ -93,13 +92,13 @@
         ],
         "effects": [],
         "text": {
-          "value": "Button",
+          "value": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. ",
           "defaultStyle": {
             "font": {
-              "postScriptName": "ArialMT",
-              "family": "Arial"
+              "postScriptName": "TimesNewRomanPSMT",
+              "family": "Times New Roman"
             },
-            "fontSize": 36,
+            "fontSize": 20,
             "lineHeight": 0,
             "letterSpacing": 0,
             "fills": [
@@ -108,9 +107,9 @@
                 "visible": true,
                 "blendMode": "NORMAL",
                 "color": {
-                  "r": 1,
-                  "g": 1,
-                  "b": 1,
+                  "r": 0,
+                  "g": 0,
+                  "b": 0,
                   "a": 1
                 }
               }
@@ -118,13 +117,13 @@
             "strokes": []
           },
           "styles": [],
-          "horizontalAlign": "CENTER",
-          "verticalAlign": "CENTER",
+          "horizontalAlign": "LEFT",
+          "verticalAlign": "TOP",
           "frame": {
             "mode": "FIXED",
             "size": {
-              "width": 250,
-              "height": 50
+              "width": 750,
+              "height": 150
             }
           },
           "baselinePolicy": "OFFSET_ASCENDER"
