+++ packages/octopus-fig/test/integration/assets/aoZAhPTVzGi1rXnU35YPVb/expected/octopus-1-3.json
@@ -1,22 +1,22 @@
 {
-  "id": "318-13",
+  "id": "1-3",
   "type": "ARTBOARD",
-  "version": "3.0.0-rc.6",
+  "version": "3.0.0-rc.19",
   "dimensions": {
     "width": 300,
     "height": 100
   },
   "content": {
-    "id": "318-13-Background",
-    "name": "Property 1=Default",
+    "id": "1-3-Background",
+    "name": "BUTTON COMP",
     "type": "MASK_GROUP",
     "visible": true,
     "opacity": 1,
     "blendMode": "PASS_THROUGH",
-    "maskBasis": "SOLID",
+    "maskBasis": "BODY_EMBED",
     "mask": {
-      "id": "318-13-BackgroundMask",
-      "name": "Property 1=Default",
+      "id": "1-3-BackgroundMask",
+      "name": "BUTTON COMP",
       "type": "SHAPE",
       "visible": true,
       "opacity": 1,
@@ -48,7 +48,7 @@
             "y0": 0,
             "y1": 100
           },
-          "cornerRadius": 40
+          "cornerRadius": 30
         },
         "fillRule": "NON_ZERO",
         "fills": [
@@ -57,37 +57,14 @@
             "visible": true,
             "blendMode": "NORMAL",
             "color": {
-              "r": 0.4062,
-              "g": 0.625,
-              "b": 0.3516,
+              "r": 0.3958,
+              "g": 0.565,
+              "b": 1,
               "a": 1
             }
           }
         ],
-        "strokes": [
-          {
-            "style": "SOLID",
-            "dashing": [],
-            "visible": true,
-            "fill": {
-              "type": "COLOR",
-              "visible": true,
-              "blendMode": "NORMAL",
-              "color": {
-                "r": 0,
-                "g": 0,
-                "b": 0,
-                "a": 1
-              }
-            },
-            "thickness": 6,
-            "position": "CENTER",
-            "lineJoin": "MITER",
-            "lineCap": "BUTT",
-            "miterLimit": 28.96,
-            "fillRule": "NON_ZERO"
-          }
-        ]
+        "strokes": []
       }
     },
     "transform": [
@@ -100,7 +77,7 @@
     ],
     "layers": [
       {
-        "id": "318-15",
+        "id": "2-4",
         "name": "Hello",
         "type": "TEXT",
         "visible": true,
@@ -119,13 +96,12 @@
           "value": "Button",
           "defaultStyle": {
             "font": {
-              "postScriptName": "Arial-BoldMT",
+              "postScriptName": "ArialMT",
               "family": "Arial"
             },
             "fontSize": 36,
             "lineHeight": 0,
             "letterSpacing": 0,
-            "letterCase": "UPPERCASE",
             "fills": [
               {
                 "type": "COLOR",
