+++ packages/octopus-fig/test/integration/assets/aoZAhPTVzGi1rXnU35YPVb/expected/octopus-182-42.json
@@ -1,7 +1,7 @@
 {
   "id": "182-42",
   "type": "ARTBOARD",
-  "version": "3.0.0-rc.6",
+  "version": "3.0.0-rc.19",
   "dimensions": {
     "width": 4382,
     "height": 3719
