+++ packages/octopus-fig/test/integration/assets/aoZAhPTVzGi1rXnU35YPVb/expected/octopus-377-19.json
@@ -1,7 +1,7 @@
 {
   "id": "377-19",
   "type": "ARTBOARD",
-  "version": "3.0.0-rc.6",
+  "version": "3.0.0-rc.19",
   "dimensions": {
     "width": 741,
     "height": 380
