+++ packages/octopus-fig/test/integration/assets/aoZAhPTVzGi1rXnU35YPVb/expected/octopus-41-3.json
@@ -1,7 +1,7 @@
 {
   "id": "41-3",
   "type": "ARTBOARD",
-  "version": "3.0.0-rc.6",
+  "version": "3.0.0-rc.19",
   "dimensions": {
     "width": 500,
     "height": 500
