+++ packages/octopus-fig/test/integration/assets/aoZAhPTVzGi1rXnU35YPVb/expected/octopus-66-3.json
@@ -1,22 +1,22 @@
 {
-  "id": "1-2",
+  "id": "66-3",
   "type": "ARTBOARD",
-  "version": "3.0.0-rc.6",
+  "version": "3.0.0-rc.19",
   "dimensions": {
-    "width": 375,
-    "height": 250
+    "width": 300,
+    "height": 500
   },
   "content": {
-    "id": "1-2-Background",
-    "name": "Frame IO",
+    "id": "66-3-Background",
+    "name": "bug-linear-gradient",
     "type": "MASK_GROUP",
     "visible": true,
     "opacity": 1,
     "blendMode": "PASS_THROUGH",
     "maskBasis": "BODY_EMBED",
     "mask": {
-      "id": "1-2-BackgroundMask",
-      "name": "Frame IO",
+      "id": "66-3-BackgroundMask",
+      "name": "bug-linear-gradient",
       "type": "SHAPE",
       "visible": true,
       "opacity": 1,
@@ -44,9 +44,9 @@
           ],
           "rectangle": {
             "x0": 0,
-            "x1": 375,
+            "x1": 300,
             "y0": 0,
-            "y1": 250
+            "y1": 500
           }
         },
         "fillRule": "NON_ZERO",
@@ -56,9 +56,9 @@
             "visible": true,
             "blendMode": "NORMAL",
             "color": {
-              "r": 1,
-              "g": 1,
-              "b": 1,
+              "r": 0.5292,
+              "g": 0.5292,
+              "b": 0.5292,
               "a": 1
             }
           }
@@ -76,19 +76,19 @@
     ],
     "layers": [
       {
-        "id": "1-3",
-        "name": "Red Rectangle",
+        "id": "66-4",
+        "name": "Rectangle 1",
         "type": "SHAPE",
         "visible": true,
         "opacity": 1,
         "blendMode": "NORMAL",
         "transform": [
+          6.123234262925839e-17,
+          -1,
           1,
-          0,
-          0,
-          1,
-          25,
-          25
+          6.123234262925839e-17,
+          50,
+          450
         ],
         "effects": [],
         "shape": {
@@ -105,7 +105,7 @@
             ],
             "rectangle": {
               "x0": 0,
-              "x1": 100,
+              "x1": 400,
               "y0": 0,
               "y1": 200
             }
@@ -113,64 +113,52 @@
           "fillRule": "NON_ZERO",
           "fills": [
             {
-              "type": "COLOR",
+              "type": "GRADIENT",
               "visible": true,
               "blendMode": "NORMAL",
-              "color": {
-                "r": 0.8917,
-                "g": 0,
-                "b": 0,
-                "a": 1
-              }
-            }
-          ],
-          "strokes": []
-        }
-      },
-      {
-        "id": "1-4",
-        "name": "Blue Ellipse",
-        "type": "SHAPE",
-        "visible": true,
-        "opacity": 1,
-        "blendMode": "NORMAL",
-        "transform": [
-          1,
-          0,
-          0,
-          1,
-          150,
-          25
-        ],
-        "effects": [],
-        "shape": {
-          "path": {
-            "type": "PATH",
-            "visible": true,
-            "transform": [
-              1,
-              0,
-              0,
-              1,
-              0,
-              0
-            ],
-            "meta": {
-              "sourceShape": "ELLIPSE"
-            },
-            "geometry": "M200,100c0,55.228 -44.772,100 -100,100c-55.2285,0 -100,-44.772 -100,-100c0,-55.2285 44.7715,-100 100,-100c55.228,0 100,44.7715 100,100z"
-          },
-          "fillRule": "NON_ZERO",
-          "fills": [
-            {
-              "type": "COLOR",
-              "visible": true,
-              "blendMode": "NORMAL",
-              "color": {
-                "r": 0.0878,
-                "g": 0.0309,
-                "b": 0.7417,
-                "a": 1
+              "gradient": {
+                "type": "LINEAR",
+                "stops": [
+                  {
+                    "color": {
+                      "r": 0,
+                      "g": 0,
+                      "b": 0,
+                      "a": 1
+                    },
+                    "position": 0.4976148307323456
+                  },
+                  {
+                    "color": {
+                      "r": 0.8417,
+                      "g": 0,
+                      "b": 0,
+                      "a": 1
+                    },
+                    "position": 0.5013185143470764
+                  },
+                  {
+                    "color": {
+                      "r": 1,
+                      "g": 1,
+                      "b": 1,
+                      "a": 1
+                    },
+                    "position": 1
+                  }
+                ]
+              },
+              "positioning": {
+                "layout": "FILL",
+                "origin": "LAYER",
+                "transform": [
+                  -399.9999523162835,
+                  200.00001192092645,
+                  -200.00001192092645,
+                  -99.99998807907087,
+                  399.99997615814067,
+                  -0.000011920927533992653
+                ]
               }
             }
           ],
