+++ packages/octopus-fig/test/integration/assets/aoZAhPTVzGi1rXnU35YPVb/expected/octopus-manifest.json
@@ -1,5 +1,5 @@
 {
-  "version": "3.0.0-rc.18",
+  "version": "3.0.0-rc.19",
   "origin": {
     "name": "FIGMA",
     "version": "0"
