+++ packages/octopus-fig/test/integration/services/tester.ts
@@ -3,7 +3,8 @@ import path from 'path'
 import * as jsondiffpatch from 'jsondiffpatch'
 
 import { createConverter } from '../../../src/index-node'
-import { getOctopusFileName, MANIFEST_NAME } from '../../../src/utils/exporter'
+import { getOctopusFileName } from '../../../src/services/exporters/node/local-exporter'
+import { MANIFEST_NAME } from '../../../src/utils/const'
 import { cleanManifest } from '../utils/asset-cleaner'
 import { stringify } from '../utils/stringify'
 import { DesignEmitterMock } from './design-emitter-mock'
